Sd1 knowledge area project step 5 review changes
