+++ app/src/main/java/com/example/twofragmentactivity/MovieDao.kt
@@ -3,6 +3,7 @@ package com.example.twofragmentactivity
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
+import androidx.room.OnConflictStrategy.REPLACE
 import androidx.room.Query
 
 @Dao
@@ -10,7 +11,7 @@ interface MovieDao {
     @Query("SELECT * FROM movies")
     fun getAllMovies(): LiveData<List<MoviesListResponse>>
 
-    @Insert
+    @Insert(onConflict = REPLACE)
     fun insertAll(movies: List<MoviesListResponse>)
 
     @Query("DELETE from movies")
