+++ app/src/main/java/com/example/twofragmentactivity/MoviesListViewModel.kt
@@ -1,13 +1,12 @@
 package com.example.twofragmentactivity
 
-import android.util.Log
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class MoviesListViewModel(var moviesDataRepository: MoviesDataRepository) : ViewModel() {
+class MoviesListViewModel( moviesDataRepository: MoviesDataRepository) : ViewModel() {
 
+    val movies : LiveData<List<MoviesListResponse>> = moviesDataRepository.getMovies()
     fun getMovieListLiveData(): LiveData<List<MoviesListResponse>> {
-        return moviesDataRepository.getMovies()
+        return movies
     }
 }
